refactor qrcode component and proving util
